@@ -18,9 +18,9 @@ const getCurrentList    = require('./getCurrentList');
 const getItemsAtRange   = require('./getItemsAtRange');
 const getPreviousItem   = require('./getPreviousItem');
 
-const KEY_ENTER     = 'enter';
-const KEY_TAB       = 'tab';
-const KEY_BACKSPACE = 'backspace';
+const KEY_ENTER     = 'Enter';
+const KEY_TAB       = 'Tab';
+const KEY_BACKSPACE = 'Backspace';
 
 /**
  * A Slate plugin to handle keyboard events in lists.
@@ -49,11 +49,10 @@ function EditList(opts = {}) {
     /**
      * User is pressing a key in the editor
      */
-    function onKeyDown(e, data, change) {
-        // Build arguments list
-        const args = [e, data, change, opts];
+    function onKeyDown(event, change, editor) {
+        const args = [event, change, editor, opts]
 
-        switch (data.key) {
+        switch (event.key) {
         case KEY_ENTER:
             return onEnter(...args);
         case KEY_TAB:
