@@ -4,7 +4,7 @@ const getCurrentItem = require('./getCurrentItem');
 /**
  * User pressed Delete in an editor
  */
-function onBackspace(event, data, change, opts) {
+function onBackspace(event, change, editor, opts) {
     const { state } = change;
     const { startOffset, selection } = state;
 
