@@ -11,10 +11,10 @@ const getItemDepth = require('./getItemDepth');
  * Enter in an empty list item should remove it
  * Shift+Enter in a list item should make a new line
  */
-function onEnter(event, data, change, opts) {
+function onEnter(event, change, editor, opts) {
     // Pressing Shift+Enter
     // should split block normally
-    if (data.isShift) {
+    if (event.shiftKey) {
         return null;
     }
 
