@@ -7,7 +7,7 @@ const getCurrentItem = require('./getCurrentItem');
  * Tab       -> Increase item depth if inside a list item
  * Shift+Tab -> Decrease item depth if inside a list item
  */
-function onTab(event, data, change, opts) {
+function onTab(event, change, editor, opts) {
     const { state } = change;
     const { isCollapsed } = state;
 
@@ -16,7 +16,7 @@ function onTab(event, data, change, opts) {
     }
 
     // Shift+tab reduce depth
-    if (data.isShift) {
+    if (event.shiftKey) {
         event.preventDefault();
 
         return decreaseItemDepth(
