@@ -8,10 +8,11 @@ module.exports = function(plugin, change) {
     plugin.onKeyDown(
         {
             preventDefault: () => {},
-            stopPropagation: () => {}
+            stopPropagation: () => {},
+            key: 'Backspace'
         },
-        { key: 'backspace' },
-        change
+        change,
+        {}
     );
 
     // Selection check
