@@ -7,9 +7,10 @@ module.exports = function(plugin, change) {
     return plugin.onKeyDown(
         {
             preventDefault: () => {},
-            stopPropagation: () => {}
+            stopPropagation: () => {},
+            key: 'Backspace'
         },
-        { key: 'backspace' },
-        change
+        change,
+        {}
     );
 };
