@@ -3,9 +3,10 @@ module.exports = function(plugin, change) {
     return plugin.onKeyDown(
         {
             preventDefault: () => {},
-            stopPropagation: () => {}
+            stopPropagation: () => {},
+            key: 'Enter'
         },
-        { key: 'enter' },
-        change
+        change,
+        {}
     );
 };
